Fix bug 35347 : [OS] Error when open the generated ps

+++ engine/org.eclipse.birt.report.engine.emitter.postscript/src/org/eclipse/birt/report/engine/emitter/postscript/truetypefont/TrueTypeFont.java
@@ -1213,7 +1213,7 @@ public class TrueTypeFont
 
 	public int getGlyphIndex( char c )
 	{
-		int[] glyphIndexs = (int[]) getCMap( ).get( c );
+		int[] glyphIndexs = (int[]) getCMap( ).get( (int) c );
 		return glyphIndexs[0];
 	}
